requirements.txt
streamlit>=1.35.0
ase>=3.22.1
mace-torch>=0.3.6
torch>=2.0.0
plotly>=5.17.0
pandas>=2.0.0
numpy>=1.24.0
